addressing comments from PR Review
